c++ array creation

+++ spigs/src/PigsDAQ.h
@@ -13,7 +13,6 @@
 #define PIGSDAQ_H_
 
 #include <stdio.h>          // printf, scanf, NULL */
-#include <stdlib.h>         // calloc, exit, free
 #include <iostream>         // std::cerr
 #include <typeinfo>         // operator typeid
 #include <exception>        // std::exception
@@ -70,12 +69,12 @@ public:
     int32_t ConfigureBoard();       // Configures the board with digitizer and DAQ parameters
     int32_t EndLibrary();           // Cleanup: call this before the program ends!
 
-    int32_t isChannelDisabled(int32_t ch);          // 0 if enabled, 1 if disabled
-    void PrintChannelParameters(int32_t ch);        // Prints Channel Parameters
+    int32_t isChannelDisabled(int8_t ch);          // 0 if enabled, 1 if disabled
+    void PrintChannelParameters(int8_t ch);        // Prints Channel Parameters
     char * decodeError(char *dest, int32_t code);   // Decodes the given error code into a message
 
-    int32_t ConfigureChannel(int32_t ch);   // Sets channel parameters specified in Init & Set calls
-    int32_t StopAcquisition(int32_t ch);    // Stops acquisition for channel ch
+    int32_t ConfigureChannel(int8_t ch);   // Sets channel parameters specified in Init & Set calls
+    int32_t StopAcquisition(int8_t ch);    // Stops acquisition for channel ch
     int32_t AcquisitionSingleLoop();        // Runs acquisition
     int32_t ThreadAcqSingleLoop();          // Runs acquisition in a separate thread
     void PrintAcquisotionInfo();            // Prints real/dead time, cps, ...
@@ -105,7 +104,7 @@ protected:
     ~PigsDAQ(){};
 
 private:
-    int32_t ch;         // ch is the variable used to identify a board's channel inside DPPLibrary.
+    int8_t ch;         // ch is the variable used to identify a board's channel inside DPPLibrary.
     int32_t handle;     // Handler to the Library
     int32_t brd;        // Board Identifier - this code assumes we only have one board
     CAENDPP_ConnectionParams_t connParam;   // Connection Parameters - Used to connect to the board.
